% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_rate.R
\name{get_historical_rate}
\alias{get_historical_rate}
\title{Get the historical currency exchange rate on a specified date}
\usage{
get_historical_rate(date, symbol = "CAD", base_symbol = "USD", access_key)
}
\arguments{
\item{date}{the date of interest, a string in the format 'YYYY-MM-DD'}

\item{symbol}{currency symbol of interest, for example "CAD"}

\item{base_symbol}{currency symbol as a reference base 1, for example "USD"}

\item{access_key}{access key for the fixer.io api, a string.}
}
\value{
currency exchange rate on the specified date
}
\description{
Get the historical currency exchange rate on a specified date
}
\examples{
get_historical_rate('2018-01-01', 'CAD', 'EUR', access_key = Sys.getenv('ACCESS_KEY'))
}
