% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_period_rate.R
\name{get_period_rate}
\alias{get_period_rate}
\title{Get currency exchange rates for a specified period}
\usage{
get_period_rate(start_date, end_date, symbol = "CAD", base_symbol = "USD",
  access_key)
}
\arguments{
\item{start_date}{the start date of interest, a string in the format 'YYYY-MM-DD'}

\item{end_date}{the end date of interest, a string in the format 'YYYY-MM-DD'}

\item{symbol}{currency symbol of interest, for example "CAD"}

\item{base_symbol}{currency symbol as a reference base 1, for example "USD"}

\item{access_key}{access key for the fixer.io api, a string}
}
\value{
a data frame of the currency exchange rates in the specified period
}
\description{
Get daily historical currency exchange rates between two dates of your choice, with a maximum time frame of 30 days.
}
\examples{
get_period_rate('2018-01-14', '2018-01-15', 'CAD', 'EUR', access_key = Sys.getenv('ACCESS_KEY'))
}
